// src/generated/source-data.ts
export const constants = {
  "KNPLAY": 10,
  "KNBASE": 10,
  "KGALV": 75,
  "KSID": 25,
  "KGALH": 75,
  "BRDSIZ": 1875,
  "KNPLNT": 60,
  "KNDEV": 9,
  "KPASS": "*MINK",
  "KNMSG": 32,
  "MSGLEN": 17,
  "KNGRP": 7,
  "KNHIT": 64,
  "KNCMD": 33,
  "KNPCMD": 16,
  "KNPOIN": 8,
  "KNXTR": 8,
  "KSCMD": 2,
  "KMAXTK": 15,
  "KNTTY": 8,
  "KRANGE": 10,
  "KCRIT": 3000,
  "KENDAM": 25000,
  "KWAIT": 0,
  "KCMDTM": 2000,
  "KQLEN": 10,
  "MAXINT": 34359738367,
  "LOCSIZ": 200,
  "KEOL": -1,
  "KNUL": 0,
  "KINT": 1,
  "KFLT": 2,
  "KALF": 3,
  "KVPOS": 1,
  "KHPOS": 2,
  "KNTURN": 3,
  "KSPCON": 4,
  "KNTORP": 5,
  "KSHCON": 6,
  "KLFSUP": 7,
  "KSNRGY": 8,
  "KSDAM": 9,
  "KSSHPC": 10,
  "KPEDAM": 1,
  "KPEKIL": 2,
  "KPBDAM": 3,
  "KPPCAP": 4,
  "KPBBAS": 5,
  "KPRKIL": 6,
  "KNSDES": 7,
  "KNPDES": 8,
  "KDSHLD": 1,
  "KDWARP": 2,
  "KDIMP": 3,
  "KDLIFE": 4,
  "KDTORP": 5,
  "KDPHAS": 6,
  "KDCOMP": 7,
  "KDRAD": 8,
  "KNJBST": 9,
  "KJOB": 1,
  "KNAM1": 2,
  "KNAM2": 3,
  "KPPN": 4,
  "KTTYN": 5,
  "KTTYSP": 6,
  "KTTYTP": 7,
  "KJOBTM": 8,
  "KRUNTM": 9,
  "NAMBIT": 1,
  "PRTBIT": 2,
  "SHPBIT": 4,
  "BASBIT": 8,
  "PLNBIT": 16,
  "CRDBIT": 32,
  "FEDBIT": 64,
  "EMPBIT": 128,
  "ROMBIT": 256,
  "NEUBIT": 512,
  "CAPBIT": 1024,
  "ALLBIT": 2048,
  "RNGBIT": 4096,
  "CLSBIT": 8192,
  "SUMBIT": 16384,
  "LSTBIT": 32768,
  "OUTMSK": 49152,
  "IRNBIT": 1,
  "ISRBIT": 2,
  "IGMBIT": 4,
  "ORNBIT": 8,
  "PASBIT": 16,
  "KNOBIT": 32,
  "SIDMSK": 1984,
  "OBJMSK": 63,
  "GREEN": 1,
  "YELLOW": 2,
  "RED": 3,
  "SHORT": -1,
  "MEDIUM": 0,
  "LONG": 1,
  "KREL": -1,
  "KBOTH": 0,
  "KABS": 1,
  "LSTCMD": 0,
  "SUMCMD": 1,
  "BASCMD": 2,
  "PLNCMD": 3,
  "TARCMD": 4,
  "DXMPTY": 0,
  "DXFSHP": 1,
  "DXESHP": 2,
  "DXFBAS": 3,
  "DXEBAS": 4,
  "DXROM": 5,
  "DXNPLN": 6,
  "DXFPLN": 7,
  "DXEPLN": 8,
  "DXSTAR": 9,
  "DXBHOL": 10
} as const;

export const commands = [
  {
    "id": 1,
    "words": [
      "BAses",
      "     "
    ],
    "name": "BAses"
  },
  {
    "id": 2,
    "words": [
      "BUild",
      "     "
    ],
    "name": "BUild"
  },
  {
    "id": 3,
    "words": [
      "Captu",
      "re   "
    ],
    "name": "Capture"
  },
  {
    "id": 4,
    "words": [
      "DAmag",
      "es   "
    ],
    "name": "DAmages"
  },
  {
    "id": 5,
    "words": [
      "DOck ",
      "     "
    ],
    "name": "DOck"
  },
  {
    "id": 6,
    "words": [
      "Energ",
      "y    "
    ],
    "name": "Energy"
  },
  {
    "id": 7,
    "words": [
      "Gripe",
      "     "
    ],
    "name": "Gripe"
  },
  {
    "id": 8,
    "words": [
      "Help ",
      "     "
    ],
    "name": "Help"
  },
  {
    "id": 9,
    "words": [
      "Impul",
      "se   "
    ],
    "name": "Impulse"
  },
  {
    "id": 10,
    "words": [
      "List ",
      "     "
    ],
    "name": "List"
  },
  {
    "id": 11,
    "words": [
      "Move ",
      "     "
    ],
    "name": "Move"
  },
  {
    "id": 12,
    "words": [
      "News ",
      "     "
    ],
    "name": "News"
  },
  {
    "id": 13,
    "words": [
      "PHase",
      "rs   "
    ],
    "name": "PHasers"
  },
  {
    "id": 14,
    "words": [
      "PLane",
      "ts   "
    ],
    "name": "PLanets"
  },
  {
    "id": 15,
    "words": [
      "POint",
      "s    "
    ],
    "name": "POints"
  },
  {
    "id": 16,
    "words": [
      "Quit ",
      "     "
    ],
    "name": "Quit"
  },
  {
    "id": 17,
    "words": [
      "RAdio",
      "     "
    ],
    "name": "RAdio"
  },
  {
    "id": 18,
    "words": [
      "REpai",
      "r    "
    ],
    "name": "REpair"
  },
  {
    "id": 19,
    "words": [
      "SCan ",
      "     "
    ],
    "name": "SCan"
  },
  {
    "id": 20,
    "words": [
      "SEt  ",
      "     "
    ],
    "name": "SEt"
  },
  {
    "id": 21,
    "words": [
      "SHiel",
      "ds   "
    ],
    "name": "SHields"
  },
  {
    "id": 22,
    "words": [
      "SRsca",
      "n    "
    ],
    "name": "SRscan"
  },
  {
    "id": 23,
    "words": [
      "STatu",
      "s    "
    ],
    "name": "STatus"
  },
  {
    "id": 24,
    "words": [
      "SUmma",
      "ry   "
    ],
    "name": "SUmmary"
  },
  {
    "id": 25,
    "words": [
      "TArge",
      "ts   "
    ],
    "name": "TArgets"
  },
  {
    "id": 26,
    "words": [
      "TEll ",
      "     "
    ],
    "name": "TEll"
  },
  {
    "id": 27,
    "words": [
      "TIme ",
      "     "
    ],
    "name": "TIme"
  },
  {
    "id": 28,
    "words": [
      "TOrpe",
      "dos  "
    ],
    "name": "TOrpedos"
  },
  {
    "id": 29,
    "words": [
      "TRact",
      "or   "
    ],
    "name": "TRactor"
  },
  {
    "id": 30,
    "words": [
      "TYpe ",
      "     "
    ],
    "name": "TYpe"
  },
  {
    "id": 31,
    "words": [
      "Users",
      "     "
    ],
    "name": "Users"
  },
  {
    "id": 32,
    "words": [
      "*Debu",
      "g    "
    ],
    "name": "*Debug"
  },
  {
    "id": 33,
    "words": [
      "*Pass",
      "word "
    ],
    "name": "*Password"
  }
] as const;
// ...
export const extraHelpWords = [
  [
    "CTL-C",
    "     "
  ],
  [
    "     ",
    "     "
  ],
  [
    "INTRO",
    "     "
  ],
  [
    "HInts",
    "     "
  ],
  [
    "INput",
    "     "
  ],
  [
    "Outpu",
    "t    "
  ],
  [
    "PAuse",
    "s    "
  ],
  [
    "PRega",
    "me   "
  ]
] as const;
// ...
export const helpText = [
  {
    "text": "\r\nYou cannot get HELP while under\r\nRED alert!\r\n",
    "file": "WARMAC.MAC",
    "line": 5018
  },
  {
    "text": "*",
    "file": "WARMAC.MAC",
    "line": 5041
  },
  {
    "text": "I don't know the term ",
    "file": "WARMAC.MAC",
    "line": 5056
  },
  {
    "text": "\r\nFor a list of commands type HELP *\r\nFor help on a particular command type HELP command\r\n\r\nBesides commands, help is also available for:\r\n\r\n",
    "file": "WARMAC.MAC",
    "line": 5082
  },
  {
    "text": "\r\nUpper case letters mark the shortest acceptable abbreviation.\r\n\r\n",
    "file": "WARMAC.MAC",
    "line": 5089
  },
  {
    "text": "Commands are:\r\n",
    "file": "WARMAC.MAC",
    "line": 5100
  },
  {
    "text": "%Can't read help file\r\n",
    "file": "WARMAC.MAC",
    "line": 5125
  },
  {
    "text": "%Can't find help on ",
    "file": "WARMAC.MAC",
    "line": 5172
  },
  {
    "text": " is ambiguous.",
    "file": "WARMAC.MAC",
    "line": 5229
  },
  {
    "text": "  Could be:\r\n",
    "file": "WARMAC.MAC",
    "line": 5232
  }
] as const;